third_party/WebKit/LayoutTests/screen_orientation/screenorientation-unsupported-no-crash.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../fast/forms/resources/picker-common.js"></script>
<input type=date id=date>
<script>
test(function() {
    function popupCallback() {
        // To handle platforms (Mac) that doesn't support testing popup menu windows,
        // phrase the check so that openPicker()'s errorCallback produces the same
        // expected output when invoked.
        var popupWindowScreenOrientation = popupWindow ? popupWindow.screen.orientation : null;
        assert_equals(popupWindowScreenOrientation, null);
    }
    openPicker(document.getElementById('date'), popupCallback, popupCallback);
}, "Verify that accessing screen.orientation where unsupported doesn't crash.");
</script>
</body>
</html>
